+++ ak_library_management/models/library_multiple_books.py
@@ -22,6 +22,7 @@ class LibraryMultipleBooks(models.TransientModel):
     price = fields.Integer(string="price")
     bulk_books_count = fields.Integer(string="Product Count",
                                       compute="_compute_bulk_books_count")
+    product_ids = fields.Many2many(comodel_name="product.template")
 
     def create_products(self):
         """
@@ -29,12 +30,18 @@ class LibraryMultipleBooks(models.TransientModel):
         which is comma separated books names given input by user.
         """
         for book_name in self.book_names.split(','):
-            # removing spaces in to the book name
-            book_name = book_name.strip()
-            if not bool(self.env['product.template'].search([('name', '=', book_name)])):
-                self.env['product.template'].create({
+            book_name.strip()
+            if not self.env['product.template'].search([('name', '=', book_name)]):
+                products = self.env['product.template'].create({
                     'name': book_name,
                     'author': self.author_id.name
+                })
+                # link or assign created product id to this model many2many product_ids
+                self.product_ids = [(4, products.id)]
+                # send the notification to the current user for created message
+                self.env['bus.bus']._sendone(self.env.user.partner_id, 'simple_notification', {
+                    'type': 'success',
+                    'message': f"{book_name} is created as product.",
                 })
 
     def revert_changes(self):
@@ -43,33 +50,28 @@ class LibraryMultipleBooks(models.TransientModel):
         If clicked, it will delete all products created
         from the current Bulk Upload Books Record session.
         """
-        book_names_list = self.book_names.split(',')
-        self.env['product.template'].search([('name', '=', book_names_list)]).unlink()
+        self.product_ids.unlink()
 
-    @api.depends("book_names")
+    @api.depends("product_ids")
     def _compute_bulk_books_count(self):
         """
-        This function compute based on the book_names field.
+        This function compute based on the product_ids field.
         count the all books or products in the current bulk
         """
-        if self.book_names:
-            book_names_list = self.book_names.split(',')
-            self.bulk_books_count = (self.env['product.template'].
-                                     search_count([("name", "in", book_names_list)]))
-        else:
-            self.bulk_books_count = 0
+        for record in self:
+            record.bulk_books_count = len(record.product_ids) if record.product_ids else 0
 
     def bulk_books(self):
         """
         This function redirect to the product list view.
         """
-        book_names_list = self.book_names.split(',')
-        product_rec = self.env['product.template'].search([("name","in",book_names_list)])
-        return {
+        action = {
             'name': 'Bulk Books',
             'type': 'ir.actions.act_window',
             'res_model': 'product.template',
-            'view_mode': 'list,form' if len(book_names_list) > 1 else 'form',
-            'domain': [("name", "in", book_names_list)] if len(book_names_list) else [],
-            'res_id': product_rec[0].id if len(product_rec) == 1 else None,
+            'view_mode': 'list,form' if len(self.product_ids) > 1 else 'form',
+            'domain': [("id", "in", self.product_ids.ids)],
         }
+        if len(self.product_ids) == 1:
+            action['res_id'] = self.product_ids.id
+        return action